requirements.txt
Flask
requests
Pillow
pytesseract
opencv-python
torch
google-generativeai
markdown2
python-dotenv